bb/requirements.txt
# Banco de Dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Ambiente
python-dotenv==1.0.0

# Autenticação
pyjwt==2.8.0

# FastAPI e Servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Validação (incluído com FastAPI mas garantindo versão)
pydantic==2.5.3
pydantic[email]==2.5.3

# CORS
python-multipart==0.0.6